Completely restructure the views
