+++ ui/src/views/auth/github/callback/index.tsx
