+++ ui/src/views/renders/[id]/RenderControls.tsx
