+++ ui/src/views/renders/[id]/RenderDisplay/index.tsx
