+++ ui/src/views/renders/new/Controls/NewTextureSpeedDial.tsx
