+++ ui/src/views/renders/new/Controls/cards/ControlsCardCamera.tsx
