+++ ui/src/views/renders/new/Controls/cards/ControlsCardTexture/index.tsx
@@ -1,7 +1,7 @@
 import { ControlsCard } from '../ControlsCard';
 import { TextInputControl } from '../form-controls/TextInputControl';
 import { TextArrayInputControl } from '../form-controls/TextArrayInputControl';
-import { InfoIconAdditionalInfo } from '@/views/render-new/Controls/icons/InfoIconAdditionalInfo';
+import { InfoIconAdditionalInfo } from '@/views/renders/new/Controls/icons/InfoIconAdditionalInfo';
 import { NestedTextureHeader } from './NestedTextureHeader';
 import { getTextureData } from '@/utils/render/texture';
 import { fixReferences } from '@/utils/render/utils';
