+++ ui/src/views/renders/new/Controls/icons/WarningIconAdvancedProperty.tsx
