+++ ui/src/views/renders/new/Controls/icons/WarningIconUnaffectedPreview.tsx
